+++ src/pages/library/DataBrowser.js
@@ -1,12 +1,43 @@
 import { h } from 'react-hyperscript-helpers'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { libraryTopMatter } from 'src/components/library-common'
+import * as StateHistory from "src/libs/state-history";
+import {SearchAndFilterTopLevelComponent} from "src/pages/library/common";
+import {useState} from "react";
+
+
+
+// Description of the structure of the sidebar. Case is preserved when rendering but all matching is case-insensitive.
+const sidebarSections = [{
+  name: 'Test',
+  labels: ['test1', 'test2'],
+  keepCollapsed: true
+}]
 
 
 const DataBrowser = () => {
-  return h(FooterWrapper, { alwaysShow: true }, [
-    libraryTopMatter('browse & explore')
-  ])
+
+  const featuredList = [{
+      namespace: "test-test",
+      name: "test1",
+      created: "2020-01-13T18:25:28.340Z",
+      tags: {
+        itemsType: "AttributeValue",
+        items: [
+          "test1",
+        ]
+      },
+      description: "test desc",
+      lowerName: "test1",
+      lowerDescription: "test desc",
+
+    keepCollapsed: true
+  }]
+
+  const [workspacesByTag, setWorkspacesByTag] = useState({})
+  const [sections, setSections] = useState([])
+
+  return SearchAndFilterTopLevelComponent(featuredList, workspacesByTag, sections, sidebarSections)
 }
 
 export const navPaths = [{
