+++ src/pages/library/Showcase.js
@@ -16,7 +16,15 @@ import * as Nav from 'src/libs/nav'
 import * as StateHistory from 'src/libs/state-history'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
-import {groupByFeaturedTags, NavItem, Pill, Sidebar, SidebarCollapser, styles} from "src/pages/library/common";
+import {
+  groupByFeaturedTags,
+  NavItem,
+  Pill,
+  SearchAndFilterTopLevelComponent,
+  Sidebar,
+  SidebarCollapser,
+  styles
+} from "src/pages/library/common";
 
 
 // Description of the structure of the sidebar. Case is preserved when rendering but all matching is case-insensitive.
@@ -49,57 +57,12 @@ const sidebarSections = [{
   labels: ['HCA', 'AnVIL', 'BRAIN Initiative', 'BioData Catalyst', 'NCI']
 }]
 
-const makeCard = variant => workspace => {
-  const { namespace, name, created, description } = workspace
-  return a({
-    href: Nav.getLink('workspace-dashboard', { namespace, name }),
-    key: `${namespace}:${name}`,
-    style: {
-      backgroundColor: 'white',
-      height: 175,
-      borderRadius: 5,
-      display: 'flex',
-      marginBottom: '1rem',
-      boxShadow: Style.standardShadow
-    }
-  }, [
-    div({
-      style: {
-        backgroundRepeat: 'no-repeat', backgroundPosition: 'center', backgroundSize: 'auto 100%', borderRadius: '5px 0 0 5px',
-        width: 87,
-        ...Utils.cond(
-          [name.toLowerCase().includes('covid'), () => ({ backgroundImage: `url(${covidBg})` })],
-          [variant === 'gatk', () => ({ backgroundColor: '#333', backgroundImage: `url(${gatkLogo})`, backgroundSize: undefined })],
-          () => ({ backgroundImage: `url(${featuredBg})`, opacity: 0.75 })
-        )
-      }
-    }),
-    div({ style: { flex: 1, minWidth: 0, padding: '15px 20px', overflow: 'hidden' } }, [
-      div({ style: { display: 'flex' } }, [
-        div({ style: { flex: 1, color: colors.accent(), fontSize: 16, lineHeight: '20px', height: 40, marginBottom: 7 } }, [name]),
-        created && div([Utils.makeStandardDate(created)])
-      ]),
-      div({ style: { lineHeight: '20px', height: 100, whiteSpace: 'pre-wrap', overflow: 'hidden' } }, [description])
-      // h(MarkdownViewer, [description]) // TODO: should we render this as markdown?
-    ])
-  ])
-}
-
 const Showcase = () => {
   const stateHistory = StateHistory.get()
   const [featuredList, setFeaturedList] = useState(stateHistory.featuredList)
+
   const [workspacesByTag, setWorkspacesByTag] = useState({})
   const [sections, setSections] = useState([])
-
-  const [selectedSections, setSelectedSections] = useState([])
-  const [selectedTags, setSelectedTags] = useState([])
-  const [searchFilter, setSearchFilter] = useState()
-  const [sort, setSort] = useState('most recent')
-
-  const sortWorkspaces = Utils.switchCase(sort,
-    ['most recent', () => _.orderBy(['created'], ['desc'])],
-    ['alphabetical', () => _.orderBy(w => _.toLower(_.trim(w.name)), ['asc'])]
-  )
 
   Utils.useOnMount(() => {
     const loadData = async () => {
@@ -140,91 +103,7 @@ const Showcase = () => {
     loadData()
   })
 
-  const filterBySections = workspaces => {
-    if (_.isEmpty(selectedSections)) {
-      return workspaces
-    } else {
-      const tags = _.uniq(_.flatMap(section => section.tags, selectedSections))
-      return _.uniq(_.flatMap(tag => workspacesByTag[tag], tags))
-    }
-  }
-  // eslint-disable-next-line lodash-fp/no-single-composition
-  const filterByTags = workspaces => _.flow(_.map(tag => _.intersection(workspacesByTag[tag]), selectedTags))(workspaces)
-  const filterByText = workspaces => {
-    const lowerSearch = _.toLower(searchFilter)
-    return _.isEmpty(lowerSearch) ?
-      workspaces :
-      _.filter(workspace => _.includes(lowerSearch, workspace.lowerName) || _.includes(lowerSearch, workspace.lowerDescription), workspaces)
-  }
-  const filteredWorkspaces = _.flow(
-    filterBySections,
-    filterByTags,
-    filterByText
-  )(featuredList)
-
-  return h(FooterWrapper, { alwaysShow: true }, [
-    libraryTopMatter('featured workspaces'),
-    !featuredList ?
-      centeredSpinner() :
-      h(Fragment, [
-        div({ style: { display: 'flex', margin: '1rem 1rem 0', alignItems: 'baseline' } }, [
-          div({ style: { width: '19rem', flex: 'none' } }, [
-            div({ style: styles.sidebarRow }, [
-              div({ style: styles.header }, 'Featured workspaces'),
-              h(Pill, {
-                count: _.size(filteredWorkspaces),
-                highlight: _.isEmpty(selectedSections) && _.isEmpty(selectedTags)
-              })
-            ]),
-            div({ style: { display: 'flex', alignItems: 'center', height: '2.5rem' } }, [
-              div({ style: { flex: 1 } }),
-              h(Link, {
-                onClick: () => {
-                  setSelectedSections([])
-                  setSelectedTags([])
-                }
-              }, ['clear'])
-            ])
-          ]),
-          h(DelayedSearchInput, {
-            style: { flex: 1, marginLeft: '1rem' },
-            'aria-label': 'Search Featured Workspaces',
-            placeholder: 'Search Name or Description',
-            value: searchFilter,
-            onChange: setSearchFilter
-          }),
-          h(IdContainer, [
-            id => h(Fragment, [
-              label({ htmlFor: id, style: { margin: '0 0.5rem 0 1rem', whiteSpace: 'nowrap' } }, ['Sort by']),
-              h(Select, {
-                id,
-                isClearable: false,
-                isSearchable: false,
-                styles: { container: old => ({ ...old, width: '10rem' }) },
-                value: sort,
-                onChange: v => setSort(v.value),
-                options: ['most recent', 'alphabetical']
-              })
-            ])
-          ])
-        ]),
-        div({ style: { display: 'flex', margin: '0 1rem' } }, [
-          div({ style: { width: '19rem', flex: 'none' } }, [
-            h(Sidebar, {
-              onSectionFilter: section => setSelectedSections(_.xor([section], selectedSections)),
-              onTagFilter: tag => setSelectedTags(_.xor([tag], selectedTags)),
-              sections,
-              selectedSections,
-              selectedTags,
-              workspacesByTag: groupByFeaturedTags(filteredWorkspaces, sidebarSections)
-            })
-          ]),
-          div({ style: { marginLeft: '1rem', minWidth: 0 } }, [
-            _.map(makeCard(), sortWorkspaces(filteredWorkspaces))
-          ])
-        ])
-      ])
-  ])
+  return SearchAndFilterTopLevelComponent(featuredList, workspacesByTag, sections, sidebarSections)
 }
 
 export const navPaths = [
